# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy API",
        "description": "API pour la gestion des plantes",
        "contact": {
            "email": "admin@blossombuddy.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/user/plants": {
            "get": {
                "tags": [
                    "UserPlant"
                ],
                "summary": "Liste les plantes de l'utilisateur connecté",
                "operationId": "0eb9a070f9e58bc55f14fa36d1fa5031",
                "responses": {
                    "200": {
                        "description": "Liste des plantes de l'utilisateur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/plant": {
            "post": {
                "tags": [
                    "UserPlant"
                ],
                "summary": "Ajoute une plante à l'utilisateur",
                "operationId": "b4273403948f66e51d9e0bbc5532127b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plant_name",
                                    "city"
                                ],
                                "properties": {
                                    "plant_name": {
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Lyon"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plante ajoutée à l'utilisateur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/plant/{id}": {
            "get": {
                "tags": [
                    "UserPlant"
                ],
                "summary": "Affiche une plante précise de l'utilisateur",
                "operationId": "896a5fb8a1bef8a548e4f5e2571b2f27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la plante de l'utilisateur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "UserPlant"
                ],
                "summary": "Supprime une plante de l'utilisateur",
                "operationId": "b8bc9546b7a4d8703ee245e3fd461fe7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plante supprimée de l'utilisateur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/plant/{id}/water": {
            "post": {
                "tags": [
                    "UserPlant"
                ],
                "summary": "Enregistre l'arrosage d'une plante de l'utilisateur",
                "operationId": "738b8224830d909741f9e815834261d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "watered_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-10-08 14:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Arrosage enregistré avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}